requirements.txt
flask
pillow
pillow-heif
python-dotenv
gunicorn
